+++ Sources/TMDB_clean_MLS/presentation/widgets/MovieRow.swift
@@ -6,75 +6,6 @@ fileprivate let formatter: DateFormatter = {
     return formatter
 }()
 
-
-struct MoviePosterImage: View {
-    var posterPath: String?
-    var posterSize: PosterSize
-
-    var body: some View {
-        if let posterPath = posterPath, let url = URL(string: "https://image.tmdb.org/t/p/w500\(posterPath)") {
-            if #available(iOS 15.0, macOS 12.0, *) {
-                AsyncImage(url: url) { phase in
-                    switch phase {
-                    case .empty:
-                        ProgressView()
-                            .frame(width: posterSize.width, height: posterSize.height)
-                    case .success(let image):
-                        image
-                            .resizable()
-                            .aspectRatio(contentMode: .fit)
-                            .frame(width: posterSize.width, height: posterSize.height)
-                            .cornerRadius(10)
-                            .shadow(radius: 5)
-                    case .failure:
-                        Image(systemName: "photo")
-                            .resizable()
-                            .aspectRatio(contentMode: .fit)
-                            .frame(width: posterSize.width, height: posterSize.height)
-                            .foregroundColor(.gray)
-                    @unknown default:
-                        EmptyView()
-                            .frame(width: posterSize.width, height: posterSize.height)
-                    }
-                }
-            } else {
-                CustomImageView(imageURL: url)
-            }
-        } else {
-            Image(systemName: "photo")
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: posterSize.width, height: posterSize.height)
-                .foregroundColor(.gray)
-        }
-    }
-
-}
-
-struct PosterSize {
-    var width: CGFloat
-    var height: CGFloat
-    
-    static let medium = PosterSize(width: 100, height: 150)
-}
-
-
-struct TitleStyle: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .font(.headline) // Set the font to headline
-            .foregroundColor(.primary) // Set the text color to primary
-            .padding(.vertical, 4) // Add vertical padding
-    }
-}
-
-extension View {
-    func titleStyle() -> some View {
-        self.modifier(TitleStyle())
-    }
-}
-
-
 @available(iOS 14, macOS 11, *)
 struct MovieRow: View {
     let movie: Movie
@@ -83,8 +14,9 @@ struct MovieRow: View {
     var body: some View {
         HStack {
             ZStack(alignment: .topLeading) {
-                MoviePosterImage(posterPath: movie.posterPath ?? "",
-                                 posterSize: .medium)
+                RemoteTMDBImage(posterPath: movie.posterPath ?? "",
+                               posterSize: .medium,
+                               image: .medium)
                     .redacted(if: movie.posterPath == nil)
             }
             .fixedSize()
